requirement.txt
speechrecognition
pyttsx3
webbrowser
pygame
requests
gtts
pyjokes
python-dotenv
